fix(identifiers): move SSA helpers out of validation to break inheritance arrow
`government/detection.rs` imported `is_itin_area` and `is_test_ssn`
from `super::validation::ssn`, reversing the detection → validation
inheritance arrow. Both functions are pure pattern-matching boolean
helpers (no `Result`, no observe deps), so they belong in a shared
module sibling to detection/ and validation/.

Created `government/common.rs` with the two helpers as `pub(super)`,
plus their tests. Switched `detection.rs:29` and `builder.rs` to
import from `super::common` instead of `super::validation`. Pruned
the now-unused re-exports from `validation/mod.rs` and the
`is_test_ssn` / `is_itin_area` test blocks from `validation/ssn.rs`.

Verified: just clippy clean, just arch-check exit 0, all 346
government-module tests pass, all 5 moved is_test_ssn tests pass
under common::tests.

Closes #158
